KYLO-2275 Upgrade action for 0.9.2 to handle searching on properties should proceed if no search plugin is available

+++ services/upgrade-service/src/main/java/com/thinkbiganalytics/metadata/upgrade/v092/IndexCategoryMetadataUpgradeAction.java
@@ -32,8 +32,6 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
-import javax.inject.Inject;
-
 /**
  * Deletes all indexed category metadata, so that it can be re-indexed with updated IDs
  */
@@ -43,7 +41,7 @@ public class IndexCategoryMetadataUpgradeAction implements UpgradeState {
 
     private static final Logger log = LoggerFactory.getLogger(IndexCategoryMetadataUpgradeAction.class);
 
-    @Inject
+    @Autowired (required = false)
     private Search searchClient;
 
     @Autowired
@@ -58,14 +56,18 @@ public class IndexCategoryMetadataUpgradeAction implements UpgradeState {
     public void upgradeTo(KyloVersion startingVersion) {
         log.info("****************** Start: Index Category Metadata Upgrade Action ****************");
 
-        String[] activeProfiles = this.environment.getActiveProfiles();
-        for (String activeProfile : activeProfiles) {
-            if (activeProfile.equals("search-esr")) {
-                int deletedDocumentCountIndexDefault = searchClient.deleteAll("kylo-categories-default", "default");
-                log.info("Total category metadata documents deleted (index: default): {}", deletedDocumentCountIndexDefault);
-                int deletedDocumentCountIndexMetadata = searchClient.deleteAll("kylo-categories-metadata", "default");
-                log.info("Total category metadata documents deleted (index: metadata): {}", deletedDocumentCountIndexMetadata);
+        if (searchClient != null) {
+            String[] activeProfiles = this.environment.getActiveProfiles();
+            for (String activeProfile : activeProfiles) {
+                if (activeProfile.equals("search-esr")) {
+                    int deletedDocumentCountIndexDefault = searchClient.deleteAll("kylo-categories-default", "default");
+                    log.info("Total category metadata documents deleted (index: default): {}", deletedDocumentCountIndexDefault);
+                    int deletedDocumentCountIndexMetadata = searchClient.deleteAll("kylo-categories-metadata", "default");
+                    log.info("Total category metadata documents deleted (index: metadata): {}", deletedDocumentCountIndexMetadata);
+                }
             }
+        } else {
+            log.info("Skipping since no search plugin is available");
         }
         log.info("****************** End: Index Category Metadata Upgrade Action ****************");
     }
